- count queued file lists in maximum number of sources for matching lists on search results - remove lists queued for matching when the bundle is removed - allow searching bundles for alternates even if auto adding of sources has been disabled
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1325 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ QueueItem.cpp
@@ -227,7 +227,7 @@ uint8_t QueueItem::getMaxSegments(int64_t filesize) const noexcept {
 #endif
 }
 
-size_t QueueItem::countOnlineUsers() const {
+int QueueItem::countOnlineUsers() const {
 	return count_if(sources.begin(), sources.end(), [](const Source& s) { return s.getUser().user->isOnline(); } );
 }
